Create discord-community-metrics
This code fetches the number of upvotes a Discord member has received in the past 24 hours by making an API call and encodes the result as a Uint256 value. It can be used for a variety of on-chain use cases, such as incentivizing positive behavior within a community or providing exclusive access to specific groups based on the number of upvotes received.